test: separate integrity checks from unit tests
