+++ tests/integrity/benchmarks/browser-bench-contract.test.ts
