+++ tests/integrity/package/component-correctness-matrix.test.ts
@@ -1,7 +1,7 @@
 import { existsSync, readFileSync } from 'node:fs';
 import { join } from 'node:path';
 import { describe, expect, it } from 'vite-plus/test';
-import { componentSurface } from './fixtures/public-surface';
+import { componentSurface } from '../../fixtures/public-surface';
 
 const UNIVERSAL_SUITES = ['a11y', 'behavior', 'determinism'] as const;
