+++ tests/integrity/package/foundations-contract.test.ts
