+++ tests/integrity/source/source-layout.test.ts
@@ -1,7 +1,7 @@
 import { existsSync, readFileSync, readdirSync } from 'node:fs';
 import { join } from 'node:path';
 import { describe, expect, it } from 'vite-plus/test';
-import { componentSurface } from './fixtures/public-surface';
+import { componentSurface } from '../../fixtures/public-surface';
 
 function readComponentDirectories() {
   return readdirSync(join(process.cwd(), 'src', 'components'), {
