+++ vitest.test.checks.config.ts
@@ -4,6 +4,6 @@ export default defineConfig({
   test: {
     environment: 'node',
     globals: true,
-    include: ['tests/unit/dev_checks/**/*.test.{ts,tsx}'],
+    include: ['tests/integrity/dev_checks/**/*.test.{ts,tsx}'],
   },
 });
